+++ src/UI/ConfirmScreen/components/TransactionSummary/TransactionSummary.tsx
@@ -1,5 +1,4 @@
 import React, { useState, MouseEvent } from 'react';
-import { LoginMethodsEnum } from '@multiversx/sdk-dapp/types/enums.types';
 
 import globals from 'assets/sass/globals.module.scss';
 import { useFormContext } from 'contexts';
@@ -9,6 +8,7 @@ import Confirm from 'UI/Confirm';
 import { NFTSFTPreview } from 'UI/NFTSFTPreview';
 
 import styles from './../confirmScreen.module.scss';
+import { getConfirmButtonLabel } from './helpers';
 
 export interface TransactionSummaryPropsType {
   isConfirmCloseBtnVisible?: boolean;
@@ -48,20 +48,10 @@ export const TransactionSummary = ({
   const [isSubmitting, setIsSubmitting] = useState(false);
 
   const isNFT = txType === TransactionTypeEnum.NonFungibleESDT;
-  let confirmText: string;
-
-  switch (providerType) {
-    case LoginMethodsEnum.walletconnect:
-    case LoginMethodsEnum.extension:
-      confirmText = 'Confirm & Check your App';
-      break;
-    case LoginMethodsEnum.ledger:
-      confirmText = 'Confirm & Check your Ledger';
-      break;
-    default:
-      confirmText = 'Confirm';
-      break;
-  }
+  const confirmButtonLabel = getConfirmButtonLabel({
+    providerType,
+    hasGuardianScreen
+  });
 
   const onCloseClick = (event: MouseEvent) => {
     event.preventDefault();
@@ -137,7 +127,7 @@ export const TransactionSummary = ({
           disabled={isSubmitting}
           onClick={onConfirmClick}
         >
-          {confirmText}
+          {confirmButtonLabel}
         </button>
 
         {isConfirmCloseBtnVisible && (