+++ src/UI/NFTSFTPreview/NFTSFTPreview.tsx
@@ -13,23 +13,33 @@ export interface NFTSFTPreviewPropsType extends PartialNftType {
 }
 
 export const NFTSFTPreview = (props: NFTSFTPreviewPropsType) => {
-  const { txType, onClick, collection, ...nft } = props;
+  const { txType, onClick, identifier, ...nft } = props;
   const { name, thumbnail } = processScamNft({
     nft
   });
+
+  const isNFTorSFT = [
+    TransactionTypeEnum.NonFungibleESDT,
+    TransactionTypeEnum.SemiFungibleESDT
+  ].includes(txType);
 
   const badge = txType === TransactionTypeEnum.NonFungibleESDT ? 'NFT' : 'SFT';
   const onPreviewClick = (event: MouseEvent) => {
     event.preventDefault();
 
     if (onClick) {
-      onClick(event, Object.assign(nft, { collection }));
+      onClick(event, Object.assign(nft, { identifier }));
     }
   };
+
+  if (!isNFTorSFT) {
+    return null;
+  }
 
   return (
     <div
       onClick={onPreviewClick}
+      data-testid='token-preview'
       className={classNames(styles.preview, {
         [styles.clickable]: Boolean(onClick)
       })}
@@ -38,8 +48,16 @@ export const NFTSFTPreview = (props: NFTSFTPreviewPropsType) => {
 
       <div className={styles.content}>
         <div className={styles.left}>
-          <div className={styles.name}>{name}</div>
-          <div className={styles.collection}>{collection}</div>
+          <div data-testid='token-preview-name' className={styles.name}>
+            {name}
+          </div>
+
+          <div
+            data-testid='token-preview-identifier'
+            className={styles.identifier}
+          >
+            {identifier}
+          </div>
         </div>
 
         <div className={styles.right}>